chore(initial): first packages - fastify
Template para uma API nodejs com fastify

+++ .eslintrc.js
@@ -30,7 +30,7 @@ module.exports = {
   'rules': {
     'indent': ['warn', 2],
     'linebreak-style': ['error', 'unix'],
-    'quotes': ['error', 'single'],
+    'quotes': ['error', 'double'],
     'semi': ['error', 'always'],
     'react/react-in-jsx-scope': 'off',
